FIX: sockets fixed 12/16
