+++ front/src/components/WebGL/class/GameManager.ts
@@ -60,8 +60,8 @@ export class GameManager {
         }
 
         ItemManager.getInstance().clearItems();
-        gameDataFromServer.itemsPos = null;
-        // WebGL 컨텍스트 해제
+        gameDataFromServer.itemsPos = [];
+                // WebGL 컨텍스트 해제
         gl.getExtension('WEBGL_lose_context')?.loseContext();
     }
