+++ front/src/components/WebGL/function/initialize.ts
@@ -1,9 +1,7 @@
 import data from '../interface/gameData';
-import {ItemManager} from "../class/ItemManager";
 import {Paddle} from '../class/Paddle';
 import {vec2} from 'gl-matrix';
 import {Ball} from '../class/Ball';
-import {Item} from "../class/Item";
 
 function initialize(state: any) {
 	if (data.canvasRef === null) return;
@@ -38,18 +36,6 @@ function initialize(state: any) {
 		data.ball.velocity = 1.5;
 		data.paddle[0].speed = 1.0;
 		data.paddle[1].speed = 1.0;
-	}
-
-	if (data.mode !== 'AI') {
-		for (let i = 0; i < 5; i++) {
-			const position = vec2.fromValues(0.0, 0.0);
-			const direction = vec2.fromValues(1.0, 0.0);
-			const velocity = 1.0;
-			const radius = 0.01;
-
-			const newItem = new Item(position, direction, velocity, radius);
-			ItemManager.getInstance().items.push(newItem);
-		}
 	}
 }
