+++ front/src/components/WebGL/function/receive.ts
@@ -2,6 +2,8 @@ import {gameDataFromServer} from "../interface/gameData";
 import {ItemManager} from "../class/ItemManager";
 import {GameManager} from "../class/GameManager";
 import data from "../interface/gameData";
+import {vec2} from 'gl-matrix'
+import {Item} from "../class/Item";
 
 function receive() {
     data.paddle[0].position[0] = gameDataFromServer.paddlePos[0][0];
@@ -15,6 +17,17 @@ function receive() {
     data.scores[1] = gameDataFromServer.scores[1];
     GameManager.scoreUpdate(null);
     if (gameDataFromServer.itemsPos && gameDataFromServer.itemsPos.length > 0) {
+        ItemManager.getInstance().clearItems();
+        for (let i = 0; i < gameDataFromServer.itemsPos.length; i++) {
+			const position = vec2.fromValues(0.0, 0.0);
+			const direction = vec2.fromValues(1.0, 0.0);
+			const velocity = 1.0;
+			const radius = 0.01;
+
+			const newItem = new Item(position, direction, velocity, radius);
+			ItemManager.getInstance().items.push(newItem);
+		}
+
         for (let i = 0; i < gameDataFromServer.itemsPos.length; i++) {
             ItemManager.getInstance().items[i].position = gameDataFromServer.itemsPos[i];
         }
