+++ front/src/routes/GameLogic.tsx
@@ -28,8 +28,6 @@ const GameLogic = () => {
 
     let socket: any = null;
     useEffect(() => {
-        if (data.endGame)
-            return;
         console.log(state.data);
         try {
             data.canvasRef = canvasRef.current;
